backend/requirements.txt
fastapi
uvicorn[standard]
motor
pydantic[email]
country-state-city
python-dotenv
